+++ CHANGELOG.md
@@ -42,6 +42,7 @@ All notable changes to this project will be documented in this file.
 - AS1/2 Hide icons panel when editing
 - AS1/2 both pcode and AS editable at the same time
 - AS3 direct editation - Allow some special words (like `override`) as identifiers
+- [#2486] AS1/2 if jump to function end as return
 
  ### Changed
 - Icon of "Deobfuscation options" menu from pile of pills to medkit